Remove chk_mod.lua
Briefly: Trivial trickery, really

In detail: Checking for lua(1) and the installed modules is a hassle

The name of this command varies per distro. It may be anything from
'lua' to 'lua51', 'lua5.1', 'lua-5.1', etc.  With the introduction
of Lua 5.2, we now may have any number of variations of those as well.

Just as with the lua(1), the modules may now come in 5.1 or 5.2 flavor,
causing the script to get pissy about the missing modules, even if
they were around.

  e.g. configure finds lua(1) (5.2) but the modules are installed for
       lua51(1) and found in a different module path, thus 'require'
       will fail

Not all modules (e.g. bitop) provide module version information, making
the module version checking impossible.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>