ard.lua: Ard.to_s: Use service internal stream ID
Since the service provides an ID (1,2,3,...), use it in
the media stream ID to make them unique. Otherwise we'll
end up with identical media stream IDs in some cases.

Use 'i' for the lack of a better prefix. This is similar to
that of youtube.lua (0.9) and the use of the "itag" value in
YouTube the media stream IDs.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>